HallucinationDetectionModel/requirements.txt
#Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
#SPDX-License-Identifier: CC-BY-NC-4.0

pytorch-lightning==1.6
rich==12.5.1
hydra_colorlog
hydra-core==1.2.0
tqdm==4.64.1
sklearn
pandas==1.4.3
torch==1.12.0  # 1.12.0+cu113
